+++ src/components/ShowLocals/index.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Filter from './Filter/Filter';
+import Filter from './Filter';
 import './component.css'
